precedence network
precedence network for the project, what needs to be done before what in general, because every feature will eventually need a touch up even after completing it.

all the issues are labeled to A to K, 
its hard to know exactly how much time it will take to do something because react is new to most of us so learning react is included in the task. So 8 ideal hours was assigned to each task.