risks of project
Here is the risk table for the project, all issues have weight in risk (what could go wrong) 
and value (the importance of the issue for the application)

Everything related to backend, database sql tables, node.js is 
high risk and high value because it is where all the pictures, accounts, comments,dates
are going to be stored for the applications. If the backend does not work, we could simulate 
it just using the front end, but the whole application would not be usable

features as in post and login are essential for feature is in comment on a picture, follow a user
to work.